Fixed last commit comments.
There was a bug that when the user chose an already revealed position the UI didnt tell him about that. that bug was fixed (added else if instead of if to one of the conditions).
Did the other TODOs.
Changed the game that the GameManager will pass the turn by himself instead of the UI.
Not finished yet, UI should not know who plays, need to fix.